Update based on the spec (RC4 CryptoAPI Encryption Header)
